requirements.txt
# Text-to-SQL MVP Dependencies
# Phase 1: Foundation

# Google AI
google-genai>=1.0.0
google-cloud-bigquery>=3.25.0
google-cloud-bigquery-storage>=2.25.0

# SQL Parsing
sqlglot>=25.0.0

# Local Database (for testing)
duckdb>=1.0.0

# Async Support
aiohttp>=3.9.0
tenacity>=8.2.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Embeddings (for dynamic few-shot)
sentence-transformers>=2.2.0

# CLI
typer>=0.9.0
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
structlog>=23.2.0
